Fix test for function move
